requirements.txt
aiohttp==1.0.5
async-timeout==3.0.0
certifi==2018.4.16
chardet==3.0.4
discord==0.0.2
discord.py==0.16.12
docopt==0.6.2
idna==2.7
monotonic==1.5
multidict==4.3.1
python-valve==0.2.1
requests==2.19.1
six==1.11.0
urllib3==1.23
websockets==3.4
